### 2020-05-28 #2235 @gkthiruvathukal
@danielskatz Just to clarify the workflow, when I am invited, if I agree, do I just assign myself as the editor, if I want to accept the invitation?
