### 2020-05-28 #2235 @danielskatz
yes, exactly - you do ```@whedon assign @gkthiruvathukal as editor```
