### 2020-06-18 #2362 @whedon
[REVIEW]: AutoFunc: A Python package for automating and verifying functional modeling
**Submitting author:** @AlexMikes (<a href="http://orcid.org/0000-0002-7873-5758">Alex Mikes</a>)
**Repository:** <a href="https://github.com/AlexMikes/AutoFunc" target ="_blank">https://github.com/AlexMikes/AutoFunc</a>
**Version:** v0.3.1
