### 2020-06-22 #2362 @AlexMikes
@cmccomb Thanks for being a reviewer and taking the time to give great feedback! I just enabled issues, should I reply here or after you've put your notes in the issue?

I'll work on adding documentation for the example files and community guidelines in the README.
