### 2020-06-22 AlexMikes/AutoFunc#1 @cmccomb
JOSS Review: Clarify contribution and authorship
Four authors are listed on the work, but only one author has made commits in the software. How did the other authors contribute to this software?

(main review issue here: https://github.com/openjournals/joss-reviews/issues/2362)
