### 2020-06-22 AlexMikes/AutoFunc#2 @cmccomb
JOSS Review: Example Usage and Functionality Documentation
There are a number of (really cool!) example scripts, but the functionality in these is largely not demonstrated or discussed in the README. Could you add a list of example scripts as a way of highlighting and documenting functionality?

(main review issue here: https://github.com/openjournals/joss-reviews/issues/2362)
