### 2020-06-22 AlexMikes/AutoFunc#3 @cmccomb
JOSS Review: Community Guidelines
Add guidance on how to report issues on problems with the software and how to seek support.

(main review issue here: https://github.com/openjournals/joss-reviews/issues/2362)
