### 2020-06-22 #2362 @cmccomb
> @AlexMikes will you please enable issues in your repository? Once you've done so I'll put my review notes in as issues. For the time being here they are:
>
> 1. Contribution and authorship: four authors are listed on the work, but only one author has made commits in the software. How did the other authors contribute to this software?
> 2. Example Usage and Functionality Documentation: There are a number of (really cool!) example scripts, but the functionality in these is largely not demonstrated or discussed in the README. Could you add a list of example scripts as a way of highlighting and documenting functionality?
> 3. Community guidelines: no guidance on how to report issues on problems with the software, or to seek support.

@AlexMikes its my pleasure! Really cool piece of software. I've added them here:
1. https://github.com/AlexMikes/AutoFunc/issues/1
2. https://github.com/AlexMikes/AutoFunc/issues/2
3. https://github.com/AlexMikes/AutoFunc/issues/3
