### 2020-07-07 AlexMikes/AutoFunc#1 @AlexMikes
The other authors provided collaboration and leadership on the functionality and direction of the software.
