### 2020-07-07 98dabb7
Update README.md
Adding a list of the examples and guidance for support
