+++ README.md
@@ -35,6 +35,8 @@ pip install pandas
 Example files are provided in the examples folder. Autofunc will automate the functional representations of components
 as  long as the format of the .csv file is has the component in column 1 and the function-flow in column 2
 
+More information on the methods used in these files can found in the various research papers that this software supports, especcially IDETC2020-22346
+"OPTIMIZING AN ALGORITHM FOR DATA MINING A DESIGN REPOSITORY TO AUTOMATE FUNCTIONAL MODELING". All of the plots for this paper were created in the ```example_optimize_with_comp_ratio.py``` file.
 
 The following lists the examples included, with their expected functionality and outputs:
 
@@ -47,6 +49,13 @@ The following lists the examples included, with their expected functionality and
 1. ```example_find_similarity.py``` will create a similarity matrix for the training dataset. This is the percent of similar components between each product in the dataset. The main diagonal of this matrix consists of ones because every product is 100% similar to itself, but the matrix is not symmetric because each product can contain a different number of components. For example, consider a case where Product 1 has 20 components and Product 2 has 40 components. If they have 10 components in common, the similarity between Product 1 and Product 2 is 10/20 = 50%, but the similarity between Product 2 and Product 1 is 10/40 = 25%. The first product of the pair is known as the ”generating” product, which is the product in the column of this matrix. This example will create a Pandas dataframe of the similarity matrix and write this to a .csv file.
 
 1. ```example_get_func_rep.py``` will create a functional representation of the components in the input file using data mining and a classification threshold. This can be used to automate functional modeling by connecting the functions and flows at the interface of components in a product. This example will write a .csv file with the results of component-function-flow and optional frequency.
+
+1. ```example_optimization.py``` incorporates all of the main modules and optimizes the similarity and classification thresholds.
+
+1. ```example_optimize_with_comp_ratio.py``` begins with ```example_optimization.py``` and also includes the stratification and optimization of a training set.
+
+1. ```example_try_best_ids.py``` is a subset of ```example_optimize_with_comp_ratio.py``` which only includes the stratified training set and some
+relevant plots
 
 
 This is the ```example_get_func_rep.py``` file:
@@ -94,6 +103,10 @@ python example_get_func_rep.py
 
 And it will generate a file ```test1.csv``` with the results of the automated functional representation of the 
  components in the ```input_file``` based on the data from the ```file_to_learn``` in the ```assets``` folder.
+ 
+ 
+## Support
+Please submit requests for support or problems with software as issues in the repository.
 
 ## Contributing
 Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
