### 2020-07-07 AlexMikes/AutoFunc#3 @AlexMikes
Added to README: [](https://github.com/AlexMikes/AutoFunc/blob/master/README.md#support)
