### 2020-07-07 AlexMikes/AutoFunc#2 @AlexMikes
Thank you, this is an excellent suggestion. I added to the README: [https://github.com/AlexMikes/AutoFunc/blob/master/README.md#usage](https://github.com/AlexMikes/AutoFunc/blob/master/README.md#usage)
