### 2020-07-07 #2362 @AlexMikes
Thank you @cmccomb for the excellent suggestions. I have completed these and responded to the issues individually.
