### 2020-07-09 #2362 @cmccomb
> Thank you @cmccomb for the excellent suggestions. I have completed these and responded to the issues individually.

Thanks @AlexMikes! My review is complete, and I recommend accept, @gkthiruvathukal .
