### 2020-07-22 #2362 @gkthiruvathukal
@openjournals/joss-eics @e-dub is having trouble to edit the review checklist (and shows as a  proper reviewer). I think we've seen this happen in the past but don't recall exactly how we resolved it.
