### 2020-10-07 #2362 @arfon
@gkthiruvathukal if at all possible we should avoid asking JOSS editors to review (@trallard and @galessiorob) as they're already busy with other JOSS work.
