### 2020-10-07 #2362 @moorepants
I'm currently reviewing another JOSS paper. Once that is done I can likely take on future ones. So I decline.
