### 2020-10-07 #2362 @gkthiruvathukal
It looks like @galessiorob is also a member of joss-editors. So we might need a few more suggestions, @AlexMikes. We'll get there soon, I hope!
