### 2020-10-13 #2362 @srmnitc
@gkthiruvathukal @AlexMikes this module looks interesting and I could find some time to help with the review. However, before I start I would like to actually comment on a major issue. It is great that the software has already been used in publications, however, I do not find any documentation except what is on the README file. I really find this very limiting and would be one blocking issue that I would raise!
