### 2020-10-13 #2362 @AlexMikes
Thanks, @srmnitc!

I agree that documentation is important. Every module has NumPy formatted docstrings that explain the functionality and all functions have docstrings that explain the functionality as well as the parameters with their expected type and purpose in the function and the returns with their type and explanation. What else would you suggest for documentation?
