### 2020-10-14 #2362 @AlexMikes
> @AlexMikes Yes, the docstrings do look very well written! I was thinking more about using [readthedocs](https://readthedocs.org/) or [github pages](https://pages.github.com/) to host the documentation. Readthedocs then would build the docs including docstrings. It would also allow you to have more details about installing, tests, examples, and so on. Although not required, you could also have interactive examples which can be set up using [mybinder](https://mybinder.org/). I would be happy to raise some issues in detail if I am the reviewer.

Thanks @srmnitc, yes please open this as an issue along with anything else you find.
