### 2020-10-20 AlexMikes/AutoFunc#9 @srmnitc
Authorship
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

Only the first author has commits on the repository. However, there are three additional authors in the paper. JOSS points out [here](https://joss.readthedocs.io/en/latest/review_criteria.html#authorship) about the authorship guidelines. What are the contributions of the other authors?
