### 2020-10-20 AlexMikes/AutoFunc#10 @srmnitc
Installation instructions
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

Although the README gives installation instructions, it is not sufficiently described. Other than the recommended installation, what about the following:

- Could one clone the repo and install?
- What versions of python are supported?
- Which operating systems can be used?
