### 2020-10-20 AlexMikes/AutoFunc#11 @srmnitc
Missing DOIs
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

There are some missing DOIs in the paper
