### 2020-10-20 AlexMikes/AutoFunc#12 @srmnitc
Example dependencies are not met
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

It is not possible to run many examples as it needs matplotlib which is not mentioned as a dependency
