### 2020-10-20 AlexMikes/AutoFunc#13 @srmnitc
Location of examples 
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

Although the examples are mentioned in the README, the location of examples folder is not correct. Both examples and tests folders should be moved to the top level of the repository.
