### 2020-10-20 AlexMikes/AutoFunc#14 @srmnitc
Automated tests
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

I can see that a tests folder is available, but there is no documentation on how to run the tests
