### 2020-10-20 AlexMikes/AutoFunc#15 @srmnitc
Community guidelines
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

JOSS mentions that the software should have [clear community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines). At the moment, it is restricted to a single line in the repository, which is not enough.
