### 2020-10-20 AlexMikes/AutoFunc#16 @srmnitc
Statement of need in documentation
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

Although there is a statement of need in the JOSS paper, there is none in the documentation
