### 2020-10-20 AlexMikes/AutoFunc#17 @srmnitc
Other software
This is part of my my review at https://github.com/openjournals/joss-reviews/issues/2362

JOSS requires to list other software or packages with comparable functionality. This is not mentioned in the JOSS paper.
