### 2020-10-20 #2362 @srmnitc
@gkthiruvathukal I have finished my first round of review and opened some issues in the repository. At the moment, it is not possible for me to check many of the boxes, but that should be possible once the issues are cleared. Overall, it seems like a well-written and useful module.

However, I have to raise some concerns regarding the scholarly effort. The module itself is fairly small, with about 1500 lines or so including the example scripts and test scripts. But the author mentions in the accompanying paper that the module has been used in publications already, which is of course wonderful and warrants publication. The DOIs for these publications are missing at the moment, once those are available I would be happy to go through them.
