### 2020-10-20 AlexMikes/AutoFunc#18 @srmnitc
General ease of use
This is part of my my review at openjournals/joss-reviews#2362

Thanks for the well-written module. As an external user who is not an expert in the field, I strongly recommend some changes in documentation. Although not blocking of acceptance in JOSS, it would help future users:

- Moving the documentation from a simple readme file to services such as readthedocs. This would ensure that the API documentation, installation instructions etc are easily accessible.

- Explanation regarding the examples: Could there a bit of background for the examples, and what do the expected results mean? One could set them up as jupyter notebooks in a way that they can be run interactively, for example, using mybinder.org

- How could one contribute or improve the repository?
