+++ paper.bib
@@ -29,6 +29,7 @@ year = {2000}
 @article{Hirtz2002,
 author = {Hirtz, Julie and Stone, Rob and McAdams, Daniel and Szykman, Simon and Wood, Kristin},
 journal = {Research in Engineering Design},
+doi = {10.1007/s00163-001-0008-3},
 pages = {65--82},
 title = {{A functional basis for engineering design: Reconciling and evolving previous efforts}},
 volume = {13},
