### 2020-10-25 AlexMikes/AutoFunc#9 @AlexMikes
From [this page](https://joss.readthedocs.io/en/latest/submitting.html#authorship):

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are.

The other authors provided active direction and collaboration on the functionality of the software, even though I was the only one who wrote and committed any code.
