### 2020-10-25 #2362 @AlexMikes
@gkthiruvathukal When I went to look at the issues from @srmnitc, I found some issues from @e-dub from September 18th. So even though he does not seem to be responding here, he still opened some issues. Has he been removed as a reviewer for a lack of communication? Most of the issues he opened are reflected in the issues from @srmnitc, can they be removed since he is not a reviewer?
