### 2020-10-25 0d59107
Update paper.md
Updating references
