### 2020-10-25 AlexMikes/AutoFunc#11 @AlexMikes
I have updated the references to include all possible DOIs. There are now three papers without DOIs (all for which this software was used). For two of them, the conference was in August but the proceedings have not yet been published by ASME (ASME-IDETC 2020). For the third without a DOI, the conference was scheduled for August and was postponed until December due to the pandemic (DCC2020).

What do you suggest I do until these DOIs are available?
