### 2020-10-25 #2362 @gkthiruvathukal
@AlexMikes Yes, but I'd make one amendment to this.  In software engineering, we generally don't delete issues. We simply mark them as resolved. What you could do is make a comment for each of these issues. Since they are duplicates (or near duplicates), it should be easy to link to the other issue in the comments field and mark it closed.  It is my understanding we don't usually remove a reviewer, especially if they did some work toward the review but perhaps experienced challenges that prevented them from finishing it. With COVID-19, many have found themselves unable to finish work. It is even taking me a lot longer to bring submissions to proper closure, so I also want to thank you for being so great during the entire process!

Once you've worked through all the issues and the two active reviewers sign off, we'll be able to move toward acceptance.
