### 2020-10-25 c3014c3
Update README.md
Adding matplotlib dependency
