### 2020-10-25 AlexMikes/AutoFunc#12 @AlexMikes
Added to the [dependencies](https://github.com/AlexMikes/AutoFunc#dependencies) section of `README.md`:

> Many of the examples also use Matplotlib for plotting. While not required to use the AutoFunc modules, it is required to run the examples. It can be installed with:

> `pip install -U matplotlib`
