### 2020-10-25 a260a10
Moving test and examples folders to top level
