+++ examples/example_get_func_rep.py
@@ -29,4 +29,4 @@ results, unmatched = get_func_rep(thresh_results, input_file, include_frequencie
 
 
 # Optional write to file - uncomment and rename to write file
-write_results_from_dict(results, 'test1.csv')
+write_results_from_dict(combos_sorted, 'func_rep1.csv')
