+++ tests/test_find_f1.py
@@ -12,7 +12,7 @@ def test1():
     """ Test to find F1 score by manually selecting product id(s) from original data to test """
 
     script_dir = os.path.dirname(__file__)
-    file_to_learn = os.path.join(script_dir, '../assets/consumer_systems.csv')
+    file_to_learn = os.path.join(script_dir, '../autofunc/assets/consumer_systems.csv')
 
     train_data = make_df(file_to_learn)
 
