+++ tests/test_get_counts.py
@@ -7,7 +7,7 @@ def test1():
     """ Testing the counter function. When using the consumer_systems dataset, housing is the most common component """
 
     script_dir = os.path.dirname(__file__)
-    file1 = os.path.join(script_dir, '../assets/consumer_systems.csv')
+    file1 = os.path.join(script_dir, '../autofunc/assets/consumer_systems.csv')
 
     df = make_df(file1)
 
