### 2020-10-25 bc6b667
Updating paths to asset folder after moving examples folder
