+++ examples/example_try_best_ids.py
@@ -29,7 +29,7 @@ start = time.time()
 
 # Dataset used for data mining
 script_dir = os.path.dirname(__file__)
-file_to_learn = os.path.join(script_dir, '../assets/consumer_systems.csv')
+file_to_learn = os.path.join(script_dir, '../autofunc/assets/consumer_systems.csv')
 
 train_data = pd.read_csv(file_to_learn)
 train_df_whole = make_df(file_to_learn)
