+++ examples/example_get_func_rep.py
@@ -10,7 +10,7 @@ import pandas as pd
 
 # Dataset used for data mining
 script_dir = os.path.dirname(__file__)
-file_to_learn = os.path.join(script_dir, '../assets/consumer_systems.csv')
+file_to_learn = os.path.join(script_dir, '../autofunc/assets/consumer_systems.csv')
 
 include_frequencies = True
 
@@ -22,7 +22,7 @@ threshold = 0.5
 thresh_results = get_top_results(combos_sorted, threshold)
 
 # Use a known product for verification
-input_file = os.path.join(script_dir, '../assets/InputExample.csv')
+input_file = os.path.join(script_dir, '../autofunc/assets/InputExample.csv')
 
 # Get dictionary of functions and flows for each component based on data mining
 results, unmatched = get_func_rep(thresh_results, input_file, include_frequencies)
