### 2020-10-25 AlexMikes/AutoFunc#13 @AlexMikes
I moved the tests and examples folders to the top level and fixed the paths to references from them
