### 2020-10-25 91ca19b
Update README.md
Updating README with info about tests
