### 2020-10-25 AlexMikes/AutoFunc#14 @AlexMikes
Added a [testing](https://github.com/AlexMikes/AutoFunc#testing) section to README.md
