### 2020-10-25 527c864
Update README.md
Updating community guidelines in README
