### 2020-10-25 AlexMikes/AutoFunc#15 @AlexMikes
Added to [Contributing](https://github.com/AlexMikes/AutoFunc#contributing) section in README:

> We welcome contributions to the autofunc package in the form of pull requests and issues made in the repository.
>
> If you are having any problems using autofunc, please open an issue. If there is some functionality you would like to see added to autofunc, you can also open an issue up to discuss that.
>
> If you have a feature that you would like to propose be integrated into autofunc, then you should open a pull request.
