### 2020-10-25 1d244ed
Update README.md
Updating installation instructions with git clone
