### 2020-10-25 890f9d9
Removing trash.py and updating examples to default to generating the similarity matrix
