+++ examples/trash.py
@@ -1,380 +1,0 @@
-"""
-
-Example showing how to implement all of the main modules and optimize the similarity and classification thresholds
-
-It performs a leave-one-out cross-validation in a triple nested loop:
-
-1. Iterate through which product is the testing set
-    2. Iterate through similarity thresholds and make training sets based on which products are within the threshold
-        3. Iterate through classification thresholds and find the F1 scores for each combination of testing product,
-        similarity threshold, and classification threshold
-
-The optimization process is simply finding the highest average F1 score and extracting the similarity and classification
-thresholds for that number.
-
-"""
-
-from autofunc.get_precision_recall import precision_recall
-from autofunc.get_top_results import get_top_results
-from autofunc.make_df import make_df
-from autofunc.find_similarities import find_similarities
-from autofunc.counter_pandas import counter_pandas
-from autofunc.split_learning_verification import split_learning_verification
-from autofunc.df_to_list import df_to_list
-import os.path
-from matplotlib import rcParams
-
-rcParams['font.family'] = 'serif'
-rcParams['font.serif'] = ['Times New Roman']
-import matplotlib.pyplot as plt
-from operator import itemgetter
-import numpy as np
-from mpl_toolkits import mplot3d
-import time
-import pandas as pd
-from statistics import mean
-
-start = time.time()
-
-# Dataset used for data mining
-script_dir = os.path.dirname(__file__)
-file_to_learn = os.path.join(script_dir, '../assets/consumer_systems.csv')
-
-train_data = pd.read_csv(file_to_learn)
-train_df_whole = make_df(file_to_learn)
-
-all_train_ids = list(map(int, train_data.id.unique()))
-
-ps_thresh = []
-f1s = []
-keep_ps = []
-keep_ps_thresh = []
-threshes = []
-ps_time = []
-comp_ratios = []
-points = []
-save_data = []
-combinations_used = []
-
-## Uncomment 1 or 2, not both
-## If generating similarity dataframe for the first time, uncomment 1 and comment 2
-## If reading in the similarity dataframe from 1, uncomment 2 and comment 1
-## Reading boolean changes datatype to string when finding keep_ids below to deal with how
-## Pandas reads in a csv to a dataframe being different than how it makes its own the first time
-
-## 1. Make similarity dataframe, takes a while so it is saved to csv first for reading in later
-# similarity_df = find_similarities(train_data)
-# similarity_df.to_csv('consumer_similarity.csv', index = True, index_label=False, header= True)
-# reading = False
-## End 1
-
-## 2. Reading in dataframe as computed above
-similarity_df = pd.read_csv('consumer_similarity.csv')
-reading = True
-## End 2
-
-
-all_comps = list(train_df_whole.comp.unique())
-num_all_comps = len(all_comps)
-
-for test_id in all_train_ids:
-    # print(test_id)
-    #
-    test_df, train_df = split_learning_verification(train_df_whole, [test_id])
-    test_list = df_to_list(test_df)
-    train_ids = list(map(int, train_df.id.unique()))
-    #
-    # Outer loop through percent similar
-    for i in range(0, 100, 10):
-
-        ps_start = time.time()
-
[…]
