### 2020-10-25 07c1e90
Update README.md
Updating README for python versions and operating systems
