### 2020-10-25 AlexMikes/AutoFunc#10 @AlexMikes
Updated [Installation](https://github.com/AlexMikes/AutoFunc#installation) section of README and added [Repository](https://github.com/AlexMikes/AutoFunc#from-repository) installation instructions
