### 2020-10-25 8053cef
Update README.md
Updating README with longer description in intro section
