### 2020-10-25 AlexMikes/AutoFunc#16 @AlexMikes
Updated [Intro](https://github.com/AlexMikes/AutoFunc#autofunc) section of README with more information
