### 2020-10-25 AlexMikes/AutoFunc#17 @AlexMikes
There is no software with comparable functionality that I know of
