### 2020-11-01 d8deaa3
Update README.md
Fixing grammar in README
