### 2020-11-01 AlexMikes/AutoFunc#18 @AlexMikes
Thanks, @srmnitc.

I realize this software is field-specific and is not immediately accessible to external users. I also realize it is challenging to evaluate software from another field with its own set of terms. From the few users I have talked to with expertise in this field, they have found the language to be intuitive, and this software to be helpful and usable. I don't intend to invest more time to substantially develop the documentation of this software. There are a lot of great resources for learning the background for the features of this module, and I have made an effort to use consistent and inclusive language so that people can find resources for any feature that they want. This software is intended for a specific group of researchers and users. Unfortunately, only one other person from that group was available for review (see closed issues from cmccomb).

Regarding the third bullet:

Added to [Contributing](https://github.com/AlexMikes/AutoFunc#contributing) section in README:

> We welcome contributions to the autofunc package in the form of pull requests and issues made in the repository.
>
> If you are having any problems using autofunc, please open an issue. If there is some functionality you would like to see added to autofunc, you can also open an issue up to discuss that.
>
> If you have a feature that you would like to propose be integrated into autofunc, then you should open a pull request.
