### 2020-11-01 #2362 @AlexMikes
Thanks, @gkthiruvathukal. I commented on each issue, linked them to a similar issue from @srmnitc (if applicable), and closed them. They are linked to your previous post if you want to check them out.
