### 2020-11-01 #2362 @AlexMikes
Thanks, @srmnitc for your thoughtful and thorough review! You had great suggestions and insight that helped improve the quality of this software package. I made comments on each issue that you opened. There is one specific issue ([Missing DOIs](https://github.com/AlexMikes/AutoFunc/issues/11)) that I would like your feedback on how you suggest I proceed. The rest of them I hope to have addressed sufficiently, but please comment where appropriate if you suggest further changes.
