### 2020-11-09 AlexMikes/AutoFunc#11 @srmnitc
Unfortunately, as I understand, if there are no DOIs available, it can then not be added to the paper. You might have to clarify with the editor.
