### 2020-11-09 AlexMikes/AutoFunc#10 @srmnitc
The installation instructions mention-
```
tested on Python 3.6+ and MacOS, Linux, and Windows operating systems.
```
However, the CI script shows
```
language: python
python:
- "3.6"
```
Where are the tests carried out for python 3.7 etc? Also for MacOS or Windows?
