### 2020-11-09 #2362 @srmnitc
@AlexMikes Thanks for your work on the issues. I have been able to check some, but others still remain unchecked. The reasons have been updated in the corresponding issues.

@gkthiruvathukal Thanks for the chance to review this submission. As I have mentioned in a previous comment, there are still two main issues that concern me.

- I am unable to verify the scholarly effort in this work as the mentioned references are without DOIs ([see here](https://github.com/AlexMikes/AutoFunc/issues/11)). From the perspective of the code, it is rather a short module, as such I still remain unable to review the scholarly effort.

- Although I am able to check some of the review boxes, the overall quality of documentation, tests are not comparable to the recent JOSS publications. This lack of documentation, in my opinion, severely limits the potential usage. The author has also mentioned that there will be no further [work in this direction](https://github.com/AlexMikes/AutoFunc/issues/10). As such no further improvements or maintenance to the code is immediately visible.

Because of these reasons, I cannot recommend the acceptance in JOSS at the moment.
