### 2020-11-09 AlexMikes/AutoFunc#11 @gkthiruvathukal
@AlexMikes and @srmnitc,  I might be missing something but I would think depositing the papers on a preprint server (e.g. ArXiv) would be an appropriate way to get a DOI for current submissions, regardless of whether they have been accepted or not.
