### 2020-11-09 #2362 @gkthiruvathukal
@srmnitc I have followed up on the DOI issue in the relevant thread. This appear to be a resolvable issue.

@AlexMikes  Can you comment on the documentation and testing issue and how you plan to address it?
