+++ .travis.yml
@@ -1,6 +1,12 @@
 language: python
 python:
 - "3.6"
+- "3.7"
+- "3.8"
+
+os:
+  - linux
+  - osx
 
 install:
   - pip install -r requirements.txt
