### 2020-11-28 0dc9344
Updating travis tests
