### 2020-11-29 AlexMikes/AutoFunc#10 @AlexMikes
Updated [Installation](https://github.com/AlexMikes/AutoFunc#installation) section of README
