### 2020-11-29 #2362 @AlexMikes
@gkthiruvathukal @srmnitc I updated the README for the [testing issue.](https://github.com/AlexMikes/AutoFunc/issues/10)

For the [missing DOIs](https://github.com/AlexMikes/AutoFunc/issues/11), I no longer have access to the LaTeX files to create a pre-print document for submission without the copyright of the conference proceedings where they are published. I can remove them from the references or can wait until the conference proceedings are published and they have DOIs.

For the [documentation issue](https://github.com/AlexMikes/AutoFunc/issues/18), each module has thorough docstrings in standard NumPy format, but I am unable to get readthedocs to autogenerate API documentation for the docstrings. I was unaware this was a requirement in order to submit a paper to JOSS.

For context, this software was used for my graduate school research. I tried submitting this paper in March but submissions were closed due to the pandemic. I graduated in June right around when the submissions finally reopened, and have been working in the industry since then. I spent a lot of time developing this software over previous years and am unable to spend any more time on it. I submitted to JOSS with the hope that others would find it useful in their research. Previous reviewers have found it useful and sufficiently documented, however, if it is the opinion of other reviewers and the editor that it is not sufficient as-is, then please withdraw my submission to JOSS and I apologize for wasting your time.
