### 2020-11-30 #2362 @gkthiruvathukal
@AlexMikes I am NOT in the camp that would like to see it withdrawn. As the handling editor, I feel like you've been responsive to my & reviewers' feedback. Please don't consider your submission a waste of time. It is not. We don't ever make any author feel unwelcome at JOSS.

Let me summarize my view:

- I am comfortable with your proposed resolution for the missing DOIs. I think, for now, it is best to remove those DOIs, if there is insufficient time to work out the issues. If you have a bit more time, another option might be to post the authors' accepted manuscript to a preprint service such as ArXiv or to ask your co-authors to post to their institutions' digital library (which would give you a DOI as well).

- It is NOT a requirement that readthedocs can autogenerate the docs. If your own Sphinx build works, that is sufficient from my point of view. (I personally host my docs using travis-sphinx and GitHub Pages (and not readthedocs).

I want to move toward acceptance on this submission and am confident that you have been responsive to feedback. I'm also confident that our remaining reviewers did a great job and are just trying to be helpful. That said, as an editor, I will make the final recommendation, especially since this is well within my areas of expertise.

Thanks for your response!
