+++ paper.bib
@@ -53,22 +53,22 @@ year = {2002}
   note = {Under Review}
 }
 
-@unpublished{edmonds2020_2,
+@inproceedings{edmonds2020_2,
   title={A Weighted Confidence Metric to Improve Automated Functional Modeling},
   author={Edmonds, Katherine and Mikes, Alex and Stone, Robert B, and DuPont, Bryony},
   journal = {Proceedings of the ASME 2020 International Design Engineering Technical Conferences and Computers and Information},
   year={2020},
-  note = {Under Review}
+  doi = {https://doi.org/10.1115/DETC2020-22495}
 }
 
 
 
-@unpublished{mikes2020,
+@inproceedings{mikes2020,
   title={Optimizing an Algorithm for Data Mining a Design Repository to Automate Functional Modeling},
   author={Mikes, Alex and Edmonds, Katherine and Stone, Robert B, and DuPont, Bryony},
   journal = {Proceedings of the ASME 2020 International Design Engineering Technical Conferences and Computers and Information},
   year={2020},
-  note = {Under Review}
+  doi = {https://doi.org/10.1115/DETC2020-22346}
 }
 
 @software{alexmikes,
