### 2020-12-02 #2362 @AlexMikes
Thanks, @gkthiruvathukal.

I checked again and the proceedings were just published for 2/3 of the papers without DOIs. I removed the third that still doesn't have one (the conference is this week).

I got sphinx-autoapi to make some API documentation, I will work on getting some more info on the hosted documentation index.
