+++ docs/conf.py
@@ -27,8 +27,8 @@ author = 'Alex Mikes'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.napoleon','autoapi.extension']
-autoapi_dirs = ['../autofunc']
+extensions = ['sphinx.ext.napoleon']
+# autoapi_dirs = ['../autofunc']
 
 
 # Add any paths that contain templates here, relative to this directory.
