+++ requirements.txt
@@ -6,6 +6,8 @@ pytest
 pytest-cov
 pandas
 apyori
+recommonmark
+sphinx-autoapi
 #
 ###### Requirements with Version Specifiers ######
 #   See https://www.python.org/dev/peps/pep-0440/#version-specifiers
