+++ docs/_sources/autoapi/autofunc/make_df/index.rst.txt
@@ -1,0 +1,34 @@
+:mod:`autofunc.make_df`
+=======================
+
+.. py:module:: autofunc.make_df
+
+.. autoapi-nested-parse::
+
+   Convert csv file into Pandas Data Frames. Removes some of the erroneous data from The Design Repository queries
+
+
+
+Module Contents
+---------------
+
+
+Functions
+~~~~~~~~~
+
+.. autoapisummary::
+
+   autofunc.make_df.make_df
+
+
+.. function:: make_df(file)
+
+   Takes a .csv file and exports a Pandas data frame
+
+   :param file: A .csv file of a SQL query
+   :type file: string
+
+   :returns: Returns a Pandas  data frame of the data in the .csv file
+   :rtype: store_data
+
+
