### 2020-12-05 5bcd7a0
Update docs after building Travis build 123 of AlexMikes/AutoFunc
The docs were built from the branch 'master' against the commit
b1bde591a054aa7dccb8decfe32dd44ec4855287.

The Travis build that generated this commit is at
https://travis-ci.org/AlexMikes/AutoFunc/jobs/747815309.

The doctr command that was run is

    /home/travis/virtualenv/python3.7.1/bin/doctr deploy . --built-docs docs/
