+++ docs/conf.py
@@ -1,69 +1,0 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
-
-
-# -- Project information -----------------------------------------------------
-
-project = 'autofunc'
-copyright = '2020, Alex Mikes'
-author = 'Alex Mikes'
-
-
-# -- General configuration ---------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = ['recommonmark','sphinx.ext.napoleon','autoapi.extension','sphinx.ext.napoleon','sphinx.ext.intersphinx']
-autoapi_dirs = ['../autofunc']
-
-napoleon_numpy_docstring = True
-
-intersphinx_mapping = {
-  'python': ('https://docs.python.org/3.6', None),
-  'pandas': ('http://pandas.pydata.org/pandas-docs/stable/', None),
-  'numpy': ('https://docs.scipy.org/doc/numpy/', None),
-}
-
-
-source_suffix = {
-  '.rst': 'restructuredtext',
-  '.txt': 'markdown',
-  '.md': 'markdown',
-}
-
-
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
-
-
-# -- Options for HTML output -------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-html_theme = 'alabaster'
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
