### 2020-12-05 77b512d
Revert "Trying Travis/GitHub pages docs"
This reverts commit edc6cd2f79661df6e51a3b72a149119b0229491a.
