### 2020-12-05 50757f5
Updating README with readthedocs link
