### 2020-12-05 AlexMikes/AutoFunc#11 @AlexMikes
I updated two of the missing DOIs and deleted the third.
