### 2020-12-05 AlexMikes/AutoFunc#18 @AlexMikes
I made API documentation and hosted it on ReadTheDocs [here](https://autofunc.readthedocs.io/en/latest/). I updated the [Usage](https://github.com/AlexMikes/AutoFunc#usage) section of the README.
