### 2021-02-02 #2362 @kyleniemeyer
Hi @AlexMikes, I'm doing some final checks before publishing - there are just a few small changes needed in the paper:

- We require a brief Statement of Need section that illustrates the research purpose of the software (part of our [required paper elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), normally checked for automatically by whedon). You can likely reuse some of the text in your summary, but add this section header.
- in your reference for the software, it looks like your name is missing a space (the `author` field should be `Alex Mikes` instead of `AlexMikes`)
- in the two 2020 ASME IDETC references (`edmonds2020_2` and `mikes2020`), you need to remove the `https://doi.org/` from the DOI fields

Please let me know when you've made these changes, and then I will publish the paper. There's no need to re-deposit into Zenodo, since we archive the paper itself. Thanks!
