+++ paper.bib
@@ -58,7 +58,7 @@ year = {2002}
   author={Edmonds, Katherine and Mikes, Alex and Stone, Robert B, and DuPont, Bryony},
   journal = {Proceedings of the ASME 2020 International Design Engineering Technical Conferences and Computers and Information},
   year={2020},
-  doi = {https://doi.org/10.1115/DETC2020-22495}
+  doi = {10.1115/DETC2020-22495}
 }
 
 
@@ -68,11 +68,11 @@ year = {2002}
   author={Mikes, Alex and Edmonds, Katherine and Stone, Robert B, and DuPont, Bryony},
   journal = {Proceedings of the ASME 2020 International Design Engineering Technical Conferences and Computers and Information},
   year={2020},
-  doi = {https://doi.org/10.1115/DETC2020-22346}
+  doi = {10.1115/DETC2020-22346}
 }
 
 @software{alexmikes,
-  author       = {AlexMikes},
+  author       = {Alex Mikes},
   title        = {{AlexMikes/AutoFunc: Automated Functional
                    Representations}},
   month        = jun,
