### 2025-06-04 #2362 @xyluo25
Hi @AlexMikes, I hope you’re doing well and enjoying a productive week. I recently submitted the pyufunc paper to JOSS (just yesterday), and I was delighted to see that JOSS recommended you as an ideal reviewer given your expertise. I know how busy schedules can be, so I would be incredibly grateful if you could kindly consider reviewing the manuscript. Your insights and feedback would mean a great deal to me and help ensure pyufunc meets the highest standards.

Please let me know if you have any questions or need additional information. Thank you so much for your time and consideration. I truly appreciate it.
